comic_summarizer/pyproject.toml
[project]
name = "comic-summarizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.31.0",
    "rarfile>=4.2",
]
